+++ worker/src/queues/workerManager.ts
@@ -1,47 +1,21 @@
 import { Job, Processor, Queue, Worker, WorkerOptions } from "bullmq";
 import {
-  BatchExportQueue,
+  getQueue,
   convertQueueNameToMetricName,
   createNewRedisInstance,
-  IngestionQueue,
-  LegacyIngestionQueue,
   logger,
   QueueName,
   recordGauge,
   recordHistogram,
   recordIncrement,
   redisQueueRetryOptions,
-  TraceUpsertQueue,
-  DatasetRunItemUpsertQueue,
-  EvalExecutionQueue,
-  ExperimentCreateQueue,
 } from "@langfuse/shared/src/server";
-import { CloudUsageMeteringQueue } from "./cloudUsageMeteringQueue";
 
 export class WorkerManager {
   private static workers: { [key: string]: Worker } = {};
 
   private static getQueue(queueName: QueueName): Queue | null {
-    switch (queueName) {
-      case QueueName.LegacyIngestionQueue:
-        return LegacyIngestionQueue.getInstance();
-      case QueueName.BatchExport:
-        return BatchExportQueue.getInstance();
-      case QueueName.CloudUsageMeteringQueue:
-        return CloudUsageMeteringQueue.getInstance();
-      case QueueName.DatasetRunItemUpsert:
-        return DatasetRunItemUpsertQueue.getInstance();
-      case QueueName.EvaluationExecution:
-        return EvalExecutionQueue.getInstance();
-      case QueueName.ExperimentCreate:
-        return ExperimentCreateQueue.getInstance();
-      case QueueName.TraceUpsert:
-        return TraceUpsertQueue.getInstance();
-      case QueueName.IngestionQueue:
-        return IngestionQueue.getInstance();
-      default:
-        throw new Error(`Queue ${queueName} not found`);
-    }
+    return getQueue(queueName);
   }
 
   private static metricWrapper(
