+++ worker/tsconfig.json
@@ -12,6 +12,5 @@
     "downlevelIteration": true,
     "resolveJsonModule": true
   },
-  "include": ["."],
   "exclude": ["node_modules", "dist", "src/**/*.test.ts"]
 }
